+++ aeron-tools/src/main/java/uk/co/real_logic/aeron/tools/RateController.java
@@ -232,14 +232,14 @@ public class RateController
 		/* The rate we _want_ to achieve, if possible.  Might not be able
 		 * to hit it exactly due to receiver pacing, etc.  But it's what
 		 * we're aiming for. */
-		private final long goalMessagesPerSecond;
+		private final double goalMessagesPerSecond;
 		/* Number of messages to send; for this interval type, this is a
 		 * hard number, not just a goal.  We _have_ to send this many
 		 * messages, no matter how long it takes or how slowly we end up
 		 * sending them. */
 		private final long messages;
 
-		protected MessagesAtMessagesPerSecondInternal(RateController rateController, long messages, long messagesPerSecond)
+		protected MessagesAtMessagesPerSecondInternal(RateController rateController, long messages, double messagesPerSecond)
 		{
 			this.rateController = rateController;
 			this.goalMessagesPerSecond = messagesPerSecond;
@@ -297,14 +297,14 @@ public class RateController
 		/* The rate we _want_ to achieve, if possible.  Might not be able
 		 * to hit it exactly due to receiver pacing, etc.  But it's what
 		 * we're aiming for. */
-		private final long goalMessagesPerSecond;
+		private final double goalMessagesPerSecond;
 		/* Number of messages to send; for this interval type, this is a
 		 * hard number, not just a goal.  We _have_ to send this many
 		 * messages, no matter how long it takes or how slowly we end up
 		 * sending them. */
 		private final double seconds;
 
-		protected SecondsAtMessagesPerSecondInternal(RateController rateController, double seconds, long messagesPerSecond)
+		protected SecondsAtMessagesPerSecondInternal(RateController rateController, double seconds, double messagesPerSecond)
 		{
 			this.rateController = rateController;
 			this.goalMessagesPerSecond = messagesPerSecond;
